move bl from home page to wallet page
